#788: Memory controls - Fixed concurrent mod
Occurred when starting sim with ROM control and subcircuit containing ROM open.
The selection models in SubGraphVerticeViewImplSelectionModel where re-created,
leading to adding them as DrawableListeners while AbstractAddressableView.bindControlView()
was running, also iterating over the same DrawableListeners.

The relevant change in SubGraphVerticeViewImplSelectionModel was introduced with #723
for reasons apparently unrelated to that issue.